feat: gateway db schema

+++ smart-buy-gateway/src/main/java/ru/petrutik/smartbuy/gateway/model/Conversation.java
@@ -2,7 +2,7 @@ package ru.petrutik.smartbuy.gateway.model;
 
 public class Conversation {
     private Long chatId;
-    private byte requestAdded;
+    private short requestAdded;
     private ConversationStatus status;
     private String clientInput;
 }
